backend/requirements.txt
fastapi
uvicorn
diffusers
transformers
torch
pillow
ngrok
nest_asyncio
